requirements.txt
amqp==5.0.6
asgiref==3.5.2
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==4.1.1
django-cleanup==5.2.0
djangorestframework==3.13.1
kombu==5.1.0
prompt-toolkit==3.0.18
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tzdata==2022.2
vine==5.0.0
wcwidth==0.2.5
